Colocate admin logic
