+++ types/admin.ts
@@ -8,9 +8,7 @@ export interface AdminControlsTypes {
       save: string;
       publish: string;
       unpublish: string;
-      drafts?: {
-        id: number;
-      }[];
+      drafts?: { id: number; notice: string }[];
     };
   };
 }
